### 2023-04-12 #204 @whedon
[PRE REVIEW]: A Data Carpentry- Style Metagenomics Workshop
**Submitting author:** @nselem (<a href="http://orcid.org/0000-0003-1697-3862">Nelly Selem</a>)
**Repository:** <a href="https://github.com/carpentries-lab/metagenomics-workshop" target ="_blank">https://github.com/carpentries-lab/metagenomics-workshop</a>
**Version:** v1.0.0
**Managing EiC:** Jordan Gorzalski
