### 2023-04-12 #204 @labarba
**Note:** This metagenomics learning module has been peer reviewed by The Carpentries in the issue: https://github.com/carpentries-lab/reviews/issues/11#issuecomment-1451841516
