### 2023-04-18 #204 @labarba
you need to accept the editor invitation first, with: `@whedon assign me as editor`
