+++ setup.md
@@ -61,7 +61,7 @@ page linked below for information about the data and access to the data files. B
 with real data, be aware that file sizes for the data are large.information 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4285900.svg)](https://doi.org/10.5281/zenodo.4285900)
 
-More information about these data will be presented in the [first episode of the Data processing and visualization for metagenomics lesson](https://carpentries-lab.github.io/metagenomics/01-background-metadata/index.html).
+More information about these data will be presented in the [first episode of the Data processing and visualization for metagenomics lesson](https://carpentries-lab.github.io/metagenomics-analysis/01-background-metadata/index.html).
 
 
 #### **Install a Bash terminal** 
@@ -110,7 +110,7 @@ These instructions assume familiarity with the command line and with installatio
 in general. There are different operating systems and many different versions
 of operating systems and environments, so these may not work on your computer. If an 
 installation doesn't work for you, please refer to the user guide for the tool listed in the table above.
-If you have difficulties with the installations or find better ways to install things in your operating system, please raise an [Issue](https://github.com/carpentries-lab/metagenomics/issues) to let us know.
+If you have difficulties with the installations or find better ways to install things in your operating system, please raise an [Issue](https://github.com/carpentries-lab/metagenomics-analysis/issues) to let us know.
 
 To make a [Conda](https://conda.io/projects/conda/en/latest/index.html) environment, first, you need to install Conda. We recommend installing the [Miniconda3](https://docs.conda.io/en/latest/miniconda.html) version. Miniconda is a package manager that includes Conda and its dependencies and simplifies the installation process. Please first install Miniconda3 (installation instructions below) and then proceed to the installation of the environment.
 
@@ -139,7 +139,7 @@ To make a [Conda](https://conda.io/projects/conda/en/latest/index.html) environm
 Once your Miniconda3 ready, follow these instructions to install and activate the metagenomics environment. 
 
 > ## Linux: Option 1 (recommended)
-> The easier way to install the environment is using the [specifications file](https://github.com/carpentries-lab/metagenomics/blob/gh-pages/files/spec-file-Ubuntu22.txt) 
+> The easier way to install the environment is using the [specifications file](https://github.com/carpentries-lab/metagenomics-analysis/blob/gh-pages/files/spec-file-Ubuntu22.txt) 
 > for Linux Ubuntu 22.04, which has the exact versions of each tool in this environment. You can use the spec file as follows:  
 > ~~~
 > $ conda create --name metagenomics --file spec-file-Ubuntu22.txt
@@ -152,7 +152,7 @@ Once your Miniconda3 ready, follow these instructions to install and activate th
 > Another way to create an environment is with a `ỳml` file.
 > This environment can be modified by adding or deleting tools in a file `metagenomics-Ubuntu22.yml`.
 > 
-> In Ubuntu 22.04, copy this file [metagenomics-Ubuntu22.yml](https://github.com/carpentries-lab/metagenomics/blob/gh-pages/files/metagenomics-Ubuntu22.yml) 
+> In Ubuntu 22.04, copy this file [metagenomics-Ubuntu22.yml](https://github.com/carpentries-lab/metagenomics-analysis/blob/gh-pages/files/metagenomics-Ubuntu22.yml) 
 > to your computer and follow the instructions in the [video tutorial](https://youtu.be/0Ks1k9gg3E0)
 {: .solution}
 
[…]
