### 2023-05-01 #204 @nselem
Thanks @JasonJWilliamsNY yes, the lesson has been thoroughly reviewed, but the short paper describing it is not. That's what I was thinking that maybe PaulaNietoG and fpsom the reviewers of the lesson could review the manuscript. Or maybe other reviewers but pointing them to the previous review link here: The Carpentries issue: https://github.com/carpentries-lab/reviews/issues/11#issuecomment-1451841516
