### 2023-06-01 #209 @whedon
[REVIEW]: A Data Carpentry- Style Metagenomics Workshop
**Submitting author:** @nselem (<a href="http://orcid.org/0000-0003-1697-3862">Nelly Selem</a>)
**Repository:** https://github.com/carpentries-lab/metagenomics-workshop
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0

- [ ] **Version:** Does the release version given match the repository release (v1.0.0)?
