### 2023-07-06 #209 @enkera
Hello @JasonJWilliamsNY. I've had a look at the workshop and it satisfies most of the criteria needed for publication. There seem to be several broken or indirect URL links, in particular the Instructor Notes, Setup,  and Metagenomics Workshop Setup page links are dead or not working properly. Other than that, everything looks good.
