+++ _extras/guide.md
@@ -114,7 +114,7 @@ they can take it from here.
 ## Installation
 
 This workshop is designed to be run on pre-imaged Amazon Web Services (AWS) instances. See the 
-[Setup page](https://carpentries-incubator.github.io/metagenomics-workshop/setup.html) for complete setup instructions. If you are
+[Setup page](https://carpentries-lab.github.io/metagenomics-workshop/setup.html) for complete setup instructions. If you are
 teaching these lessons, and would like an AWS instance to practice on, please contact [team@carpentries.org](mailto: team@carpentries.org).
 
 ## Common problems
