+++ _extras/guide.md
@@ -244,7 +244,7 @@ even a reluctant audience started.
 
 ## Reference
 
-This Instructor's Notes are based on the [corresponding page](https://datacarpentry.org/organization-genomics/guide/index.html) in the Project Organization and Management for Genomics lesson.
+This Instructor's Notes are based on the [corresponding page](https://datacarpentry.org/genomics-workshop/guide/index.html) in the Project Organization and Management for Genomics lesson.
 
 
 # Notes for the Introduction to the Command Line for Metagenomics lesson
@@ -257,7 +257,7 @@ either because there is no good alternative for the type of analysis they want t
 things manually on individual files is unfeasible. 
 
 This lesson will introduce learners to fundamental skills needed for working with their computers through a command-line interface (using
-the bash shell). They will learn how to navigate their file system, computationally manipulate their files (e.g. copying, moving, renaming), search files, redirect output and write shell scripts. By the end of the lesson, learners will be prepared to move on to using more advanced bioinformatics command line tools (see the lesson on [Data processing and visualization for metagenomics](https://carpentries-incubator.github.io/metagenomics/)).
+the bash shell). They will learn how to navigate their file system, computationally manipulate their files (e.g. copying, moving, renaming), search files, redirect output and write shell scripts. By the end of the lesson, learners will be prepared to move on to using more advanced bioinformatics command line tools (see the lesson on [Data Processing and Visualization for Metagenomics](https://carpentries-lab.github.io/metagenomics-analysis/)).
 
 ## Lesson design
 
@@ -277,7 +277,7 @@ In order to edit your command prompt, type `PS1='\W\ $ '` into your shell and pr
 
 In order to reset the command prompt, type `source .bashrc` in order to source the bash profile, or type `PS1="\u@\h:\w $ "` in order to set the prompt to show username, "@", hostname, ":", and current working directory (ie. the user's current location within the filesystem). 
 
-NOTE: Editing the prompt is discussed in [episode 01 - Introducing the Shell](https://carpentries-incubator.github.io/shell-metagenomics/01-introduction/index.html) under the 'Navigating your file system' section. This explains how to edit the prompt via `PS1='\W\ $ '` as here, so it would perhaps be best to *start* the lesson with the *default* prompt (as all the learners will and they can see that their screen will reassuringly match the instructor's screen at this point), and then instructors can choose to edit their prompt and talk through how they're doing that for learners' benefit at this section, or the instructor can just make the change early in the lesson for the visibility benefit, and explain to learners that they can find out how to do this in the lesson materials. 
+NOTE: Editing the prompt is discussed in [Introducing the Shell](https://carpentries-lab.github.io/metagenomics-shell/01-introduction/index.html) under the 'Navigating your file system' section. This explains how to edit the prompt via `PS1='\W\ $ '` as here, so it would perhaps be best to *start* the lesson with the *default* prompt (as all the learners will and they can see that their screen will reassuringly match the instructor's screen at this point), and then instructors can choose to edit their prompt and talk through how they're doing that for learners' benefit at this section, or the instructor can just make the change early in the lesson for the visibility benefit, and explain to learners that they can find out how to do this in the lesson materials. 
 
 Resetting the command prompt is not currently included in the lesson materials, so it might be useful to be familiar with this beforehand in case of learners' questions.
 
@@ -292,27 +292,27 @@ If a learner accidentally deletes or overwrites data files, you can have them ch
 
 ## Reference
 
-This Instructor notes are based on the [corresponding page](https://datacarpentry.org/shell-genomics/guide/index.html) in the Introduction to the Command Line for Genomics lesson.
[…]
