### 2023-07-10 #209 @nselem
Hello @enkera ,
We have updated Setup and instructor notes in Metagenomics Workshop, I think they broke when we change the repo from carpentries-incubator to carpentries-lab, thank you for noticing that.
