+++ README.md
@@ -35,7 +35,7 @@ A list of contributors to the lesson can be found in [AUTHORS](AUTHORS)
 
 Please cite as:
 
-Nelly Sélem Mojica; Diego Garfias Gallegos; Claudia Zirión Martínez; Jesús Abraham Avelar Rivas; Aaron Jaime Espinosa; Abel Lovaco Flores; Tania Vanessa Arellano Fernandez (2022, Jan). Metagenomics for Software Carpentry lesson, Jan 2022. Zenodo. https://doi.org/10.5281/zenodo.4285900  
+Nelly Sélem Mojica; Diego Garfias Gallegos; Claudia Zirión Martínez; Jesús Abraham Avelar Rivas; Aaron Espinosa Jaime; Abel Lovaco Flores; Tania Vanessa Arellano Fernandez (2022, Jan). Metagenomics for Software Carpentry lesson, Jan 2022. Zenodo. https://doi.org/10.5281/zenodo.4285900  
 
 ## Acknowledgments  
 We thank UNAM for funding proyecto PAPIME "Desarrollo de material didáctico de Bioinformática con énfasis en metagenómica para las modalidades presencial y virtual"
