### 2023-08-12 #209 @labarba
@openjournals/dev – On this review thread, reviewer @enkera was added before we moved JOSE to the new editorial workflow and bot, and I suspect he didn't accept the invitation in time to get the necessary permissions to edit the review auto-generated checklist. He now tried the command `@editorialbot generate my checklist` to get a new one, but it didn't work. Can you help us?
