### 2023-08-17 #209 @xuanxu
There was a problem with the bot not knowing which checklist to create (software or learning module review).
It should be fixed now.

@enkera: can you try again the `@editorialbot generate my checklist` command?
