### 2023-08-29 e31ce1a
minor editorial suggestions
