+++ paper.md
@@ -45,14 +45,14 @@ bibliography: paper.bib
 ---
 
 # Summary
-Metagenomic analyses aim to explore the genomic diversity of communities in specific habitats by processing their DNA-sequencing data. This analysis is achieved with specialized bioinformatics tools, which often require previous coding experience. Furthermore, beginners can struggle to build a pipeline from raw data to valuable biological insights. [The Carpentries](https://carpentries.org)  hosts worldwide-used open lessons to analyze specialized datasets for beginners, including [a Data Carpentry curriculum for individuals working with genomics sequencing data](https://datacarpentry.org/genomics-workshop/). However, a lesson addressing the specific challenges associated with metagenomics data and analyses was missing. We created a complete Metagenomics curriculum in [The Carpentries Incubator](https://carpentries-incubator.org), adapting and expanding on the Data Carpentry genomics curriculum. The curriculum provides an introduction to programming, teaching learners to access and handle metagenomics data, and to run commands with the software needed for completing metagenomics analyses. Content and exercises have been improved based on experience gathered in teaching the curriculum in three 16-hour online workshops. We expect to continue to enhance this lesson, which we hope is helpful as a teaching resource for new instructors in the field, and as a guide for newcomers wishing to perform metagenomic analyses from scratch.
+Metagenomic analyses aim to explore the genomic diversity of communities in specific habitats by processing their DNA sequencing data. This analysis is achieved with specialized bioinformatics tools, which often require previous coding experience. Furthermore, beginners can struggle to build a pipeline from raw data to valuable biological insights. [The Carpentries](https://carpentries.org)  hosts open lessons used worldwide to analyze specialized datasets for beginners, including [a Data Carpentry curriculum for individuals working with genomics sequencing data](https://datacarpentry.org/genomics-workshop/). However, a lesson addressing the specific challenges associated with metagenomics data and analyses was missing. We created a complete Metagenomics curriculum in [The Carpentries Incubator](https://carpentries-incubator.org), adapting and expanding on the Data Carpentry genomics curriculum. The curriculum provides an introduction to programming, teaching learners to access and handle metagenomics data, and to run commands with the software needed for completing metagenomics analyses. Content and exercises have been improved based on experience gathered in teaching the curriculum in three 16-hour online workshops. We expect to continue to enhance this lesson, which we hope is helpful as a teaching resource for new instructors in the field, and as a guide for newcomers wishing to perform metagenomic analyses from scratch.
 
 # Statement of Need
 Bioinformatic tools are now essential to our understanding of biological systems. Open lessons for general purpose coding languages and specialized topics such as genomics, ecology, and even metagenomics are already available [@darling_data:2019], [@noauthor_lessons_nodate], [@kruchten_curricular:2020-1]. Nevertheless, a complete guide for shotgun metagenomics, assuming no prior knowledge of coding and provided hardware and software solutions, was missing.  We introduce a curriculum to fill this gap, which teaches the skills required to conduct a comprehensive metagenomics workflow with Bash and R programming in a pre-installed remote server.
 
 # Content and Learning Objectives
 
-The curriculum comprises four lessons (Fig 1). It assumes no previous programming experience but expects that students understand basic concepts of molecular biology and microbiology. The first two lessons, introductions to project planning and organization and programming in Bash respectively, are adapted from equivalent lessons in the Data Carpentry Genomics curriculum [@becker_datacarpentrygenomics-workshop:2019]. The third part includes a brief introduction to R, and the fourth teaches a complete shotgun metagenomics workflow using public data [@okie_genomic:2020].
[…]
