### 2023-08-29 #209 @JasonJWilliamsNY
Hi @ @nselem I have made a few optional suggestions via a pull request https://github.com/carpentries-lab/metagenomics-workshop/pull/20

Once you are ready, the next step is for you to create an archive on either https://zenodo.org/ or https://figshare.com/ and obtain a DOI (which you'll need to give to us). I can then move to the final acceptance and publication steps.

Tip: Here is a nice guide on doing the above with Zenodo: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content#:~:text=Making%20Your%20Code%20Citable%201%20Choose%20your%20repository.,a%20shiny%20new%20badge%20showing%20your%20new%20DOI%21
