### 2023-08-29 #209 @labarba
Do note that the instructions provided in the link above result in the following behavior:

> Zenodo archives your repository and issues a new DOI each time you create a new GitHub [release](https://docs.github.com/en/repositories/releasing-projects-on-github/about-releases).

This may or may not be behavior that you want for your project. Also note that the automatic Zenodo archives are titled with your repository name, and list every committer to the repo as an author. We request that authors edit the metadata of the Zenodo deposit so title and author list match the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object."
