+++ paper.md
@@ -48,7 +48,7 @@ bibliography: paper.bib
 Metagenomic analyses aim to explore the genomic diversity of communities in specific habitats by processing their DNA sequencing data. This analysis is achieved with specialized bioinformatics tools, which often require previous coding experience. Furthermore, beginners can struggle to build a pipeline from raw data to valuable biological insights. [The Carpentries](https://carpentries.org)  hosts open lessons used worldwide to analyze specialized datasets for beginners, including [a Data Carpentry curriculum for individuals working with genomics sequencing data](https://datacarpentry.org/genomics-workshop/). However, a lesson addressing the specific challenges associated with metagenomics data and analyses was missing. We created a complete Metagenomics curriculum in [The Carpentries Incubator](https://carpentries-incubator.org), adapting and expanding on the Data Carpentry genomics curriculum. The curriculum provides an introduction to programming, teaching learners to access and handle metagenomics data, and to run commands with the software needed for completing metagenomics analyses. Content and exercises have been improved based on experience gathered in teaching the curriculum in three 16-hour online workshops. We expect to continue to enhance this lesson, which we hope is helpful as a teaching resource for new instructors in the field, and as a guide for newcomers wishing to perform metagenomic analyses from scratch.
 
 # Statement of Need
-Bioinformatic tools are now essential to our understanding of biological systems. Open lessons for general purpose coding languages and specialized topics such as genomics, ecology, and even metagenomics are already available [@darling_data:2019], [@noauthor_lessons_nodate], [@kruchten_curricular:2020-1]. Nevertheless, a complete guide for shotgun metagenomics, assuming no prior knowledge of coding and provided hardware and software solutions, was missing.  We introduce a curriculum to fill this gap, which teaches the skills required to conduct a comprehensive metagenomics workflow with Bash and R programming in a pre-installed remote server.
+Bioinformatic tools are now essential to our understanding of biological systems. Open lessons for general-purpose coding languages and specialized topics such as genomics, ecology, and even metagenomics are already available [@darling_data:2019], [@noauthor_lessons_nodate], [@kruchten_curricular:2020-1]. Nevertheless, a complete guide for shotgun metagenomics, assuming no prior knowledge of coding and provided hardware and software solutions, was missing.  We introduce a curriculum to fill this gap, which teaches the skills required to conduct a comprehensive metagenomics workflow with Bash and R programming in a pre-installed remote server.
 
 # Content and Learning Objectives
 
@@ -61,7 +61,7 @@ Fig 1. The four lessons in the curriculum for a metagenomics workshop.
 # Learning objectives
 
 Project organization and management for metagenomics
-- Plan, structure, organize and document metagenomics data, metadata, and bioinformatics workflow.
+- Plan, structure, organize, and document metagenomics data, metadata, and bioinformatics workflow.
 - Access public data on the NCBI sequence read archive (SRA) database.
 
 Introduction to the command-line for metagenomics
@@ -94,16 +94,16 @@ The workshop includes 16 hours of content with live coding, formative assessment
 1. Cloud setup: We set up remote machines to standardize the learning environment and teaching experience. This setup lowers entry barriers for students without experience in technical installations and provides enough computing power regardless of individual configuration. Students only need a computer with internet access and a terminal program installed. We also enlist the instructions for setting up the remote machines and an alternative installation guide to be used by people who prefer to follow the lessons on a different computer.
 2. Standardized episodes: Following The Carpentries instructional design, lessons consist of multiple short _episodes_ of content, each prefaced with the questions to be answered and the learning objectives for that section, and ending with concrete, clear messages of the learned content.
 3. Teaching strategies
-    - Live-coding: Programming simultaneously with students gives a practical coding experience with examples of mistakes arising and being solved.
[…]
