### 2023-08-29 #209 @Czirion
Thank you @JasonJWilliamsNY  and @labarba

We had already made these archives in Zenodo for each repository of the lesson:
Metagenomics R: https://zenodo.org/badge/latestdoi/315204217
Metagenomics Analysis: https://zenodo.org/badge/latestdoi/287395428
Metagenomics shell: https://zenodo.org/badge/latestdoi/344547242
Metagenomics organization: https://zenodo.org/badge/latestdoi/349858841
Metagenomics workshop: https://zenodo.org/badge/latestdoi/315250269

Are those what you need? Or should I make something differently?
They do have long lists of authors, I'll see if it can be edited without affecting the DOI.
