### 2023-08-29 #209 @JasonJWilliamsNY
I think you can issue a new version; we try and make sure the Zenodo and the paper authors are the same. Since this is new maybe @labarba can comment.
