### 2023-08-30 #209 @nselem
Yes I think you are right Jason, because our links conserve as authors "Raniere Silva; François Michonneau; Maxim Belkin; Nelly Sélem; Andy Boughton; Claudia Zirión Martínez; Erin Becker; Abby Cabunoc Mayes; Katrin Leinweber
@gitlabhq; W. Trevor King; Aaron O'Leary; maneesha sane; Toby Hodges; João Rodrigues; Diego Garfias Gallegos; Rayna M Harris; Jonah Duckles; Kate Hertweck; John Blischak; Trevor Bekolay; Jemma Stachelek; AbrahamAvelar; Piotr Banaszkiewicz; aaronejaime; Edward Wallace; David Mawdsley; Ivan Gonzalez; Naupaka Zimmerman; Zhian N. Kamvar; Michael R. Crusoe" and we need to keep only the authors of this lesson.

"
