### 2023-08-30 #209 @labarba
To change the list of authors, you edit the _metadata_ of the archive, which does not issue a new DOI.
