### 2023-08-31 #209 @Czirion
The contents of the lesson itself are spread out in several GitHub repositories and each of those has a Zenodo archive. The paper.md is in the landing repo of the lesson, whose Zenodo archive is https://zenodo.org/record/7713380
