### 2023-08-31 #209 @labarba
Is it possible to make one fat .zip with everything and deposit in its own Zenodo archive?
