### 2023-09-01 #209 @JasonJWilliamsNY
> :warning: [Error preparing paper acceptance](https://github.com/openjournals/jose-papers/actions/runs/6048508678?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ```yaml
> IDREFS attribute rid references an unknown ID "ref-becker_datacarpentrygenomics-workshopU003A2019"
> ```
>

@labarba any clues? I'm not at a computer now but can look into later if not resolved.
