+++ paper.bib
@@ -218,7 +218,7 @@
 
 
 
-@article{becker_datacarpentrygenomics-workshop:019,
+@article{becker_datacarpentrygenomics-workshop:2019,
 	title = {datacarpentry/genomics-workshop: Data Carpentry: Genomics Workshop Overview, June 2019},
 	rights = {Creative Commons Attribution 4.0 International, Open Access},
 	url = {https://zenodo.org/record/3260309},
