### 2023-09-01 #209 @tobyhodges
Not sure if this will address the error above, but I have just opened a PR to fix the year on the citation for DC Genomics in the `paper.bib` file https://github.com/carpentries-lab/metagenomics-workshop/pull/21
