### 2024-01-11 #209 @labarba
Hi everybody — I apologize for the long neglect. I was slammed with editorial load from another publication in the last few months, but I have _stepped down from that role_. This will allow me to come back to JOSE and attend to various neglected submissions. Thank you for your patience.

I have checked the submission repository, and do not find a tagged v1.0.0 – could you please check that?
