### 2024-01-17 #209 @nselem
Welcome back Lorena!
I have added a tag to the workshop repository
https://github.com/carpentries-lab/metagenomics-workshop/tree/v1.0.0
